requirements.txt
requests==2.31.0
pandas==2.2.0
google-cloud-storage==2.14.0
urllib3==2.0.7